Refactor initialization
